gfs2_edit: Use PRIu64 instead of casting uint64_t to unsigned long long
Also fix up some places where the wrong type is used for block addresses.

+++ gfs2/edit/savemeta.c
@@ -1131,8 +1131,8 @@ void savemeta(char *out_fn, int saveoption, int gziplevel)
 	blks_saved = 0;
 	if (sbd.gfs1)
 		sbd.sd_bsize = sbd.sd_bsize;
-	printf("There are %llu blocks of %u bytes in the filesystem.\n",
-	                     (unsigned long long)sbd.fssize, sbd.sd_bsize);
+	printf("There are %"PRIu64" blocks of %u bytes in the filesystem.\n",
+	       sbd.fssize, sbd.sd_bsize);
 
 	printf("Filesystem size: %.2fGB\n", (sbd.fssize * sbd.sd_bsize) / ((float)(1 << 30)));
 	get_journal_inode_blocks();
